refactor(rel): extract list quantifier and comprehension execution (#1189)
